Fix integration specs
