+++ spec/integration/outputs/index_spec.rb
@@ -13,7 +13,7 @@ shared_examples "an indexer" do
       pipeline = LogStash::Pipeline.new(config)
       pipeline.run
 
-      index_url = "http://#{get_host}:#{get_port('http')}/#{index}"
+      index_url = "http://#{get_host}:#{get_port}/#{index}"
 
       ftw = FTW::Agent.new
       ftw.post!("#{index_url}/_refresh")
@@ -53,7 +53,7 @@ describe "an indexer with custom index_type", :integration => true do
       output {
         elasticsearch {
           host => "#{get_host()}"
-          port => "#{get_port('http')}"
+          port => "#{get_port}"
           index => "#{index}"
           flush_size => #{flush_size}
         }
@@ -77,7 +77,7 @@ describe "an indexer with no type value set (default to logs)", :integration => 
       output {
         elasticsearch {
           host => "#{get_host()}"
-          port => "#{get_port('http')}"
+          port => "#{get_port}"
           index => "#{index}"
           flush_size => #{flush_size}
         }
